+++ GHEtool/test/general_tests/test_GHEtool.py
@@ -49,8 +49,8 @@ def test_borefield():
     borefield.set_Rb(0.2)
 
     # set temperature boundaries
-    borefield.set_max_ground_temperature(16)  # maximum temperature
-    borefield.set_min_ground_temperature(0)  # minimum temperature
+    borefield.set_max_avg_fluid_temperature(16)  # maximum temperature
+    borefield.set_min_avg_fluid_temperature(0)  # minimum temperature
 
     assert borefield.simulation_period == 20
     assert borefield.Tf_min == 0
@@ -83,8 +83,8 @@ def borefield():
     borefield.set_Rb(0.2)
 
     # set temperature boundaries
-    borefield.set_max_ground_temperature(16)  # maximum temperature
-    borefield.set_min_ground_temperature(0)  # minimum temperature
+    borefield.set_max_avg_fluid_temperature(16)  # maximum temperature
+    borefield.set_min_avg_fluid_temperature(0)  # minimum temperature
     return borefield
 
 
@@ -99,8 +99,8 @@ def borefield_custom_data():
     borefield.create_custom_dataset()
 
     # set temperature boundaries
-    borefield.set_max_ground_temperature(16)  # maximum temperature
-    borefield.set_min_ground_temperature(0)  # minimum temperature
+    borefield.set_max_avg_fluid_temperature(16)  # maximum temperature
+    borefield.set_min_avg_fluid_temperature(0)  # minimum temperature
     return borefield
 
 
@@ -138,13 +138,13 @@ def test_imbalance(borefield):
 
 def test_sizing_L3_threshold_depth_error(borefield):
     max_temp = borefield.Tf_max
-    borefield.set_max_ground_temperature(14)
+    borefield.set_max_avg_fluid_temperature(14)
     borefield.set_ground_parameters(data_ground_flux)
     borefield._calculation_setup.use_constant_Tg = False
     with raises(UnsolvableDueToTemperatureGradient):
         borefield.gfunction(3600, borefield.THRESHOLD_DEPTH_ERROR + 1)
     borefield._calculation_setup.use_constant_Tg = True
-    borefield.set_max_ground_temperature(max_temp)
+    borefield.set_max_avg_fluid_temperature(max_temp)
     borefield.ground_data.flux = 0
 
 
@@ -258,8 +258,8 @@ def test_no_ground_data():
     borefield.set_borefield(borefield_gt)
 
     # set temperature boundaries
-    borefield.set_max_ground_temperature(16)  # maximum temperature
-    borefield.set_min_ground_temperature(0)  # minimum temperature
+    borefield.set_max_avg_fluid_temperature(16)  # maximum temperature
+    borefield.set_min_avg_fluid_temperature(0)  # minimum temperature
     with raises(ValueError):
         borefield.size()